issue-109
- Modularize login
