+++ angular/src/app/app-routing.module.ts
@@ -1,13 +1,10 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
-import {LoginComponent} from './pages/login/login.component';
 
 const routes: Routes = [
-    {path: 'login', component: LoginComponent},
+    {path: 'login', loadChildren: () => import('./login/login.module').then(m => m.LoginModule)},
     {path: 'people', loadChildren: () => import('./people/people.module').then(m => m.PeopleModule)},
     {path: 'transactions', loadChildren: () => import('./transactions/transactions.module').then(m => m.TransactionsModule)},
-    // {path: 'people/:id', component: TransactionsComponent},
-    // {path: 'transaction', component: TransactionPageComponent},
 ];
 
 @NgModule({
