+++ angular/src/app/app.module.ts
@@ -4,73 +4,25 @@ import {BrowserModule} from '@angular/platform-browser';
 import {AppRoutingModule} from './app-routing.module';
 import {AppComponent} from './app.component';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import {MatButtonModule} from '@angular/material/button';
-import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
-import {MatToolbarModule} from '@angular/material/toolbar';
-import {MatIconModule} from '@angular/material/icon';
-import {MatMenuModule} from '@angular/material/menu';
-import {FlexLayoutModule} from '@angular/flex-layout';
-import {MatDividerModule} from '@angular/material/divider';
-import {LoginComponent} from './pages/login/login.component';
 import {HTTP_INTERCEPTORS, HttpClientModule} from '@angular/common/http';
-import {MatCardModule} from '@angular/material/card';
-import {MomentModule} from 'ngx-moment';
-import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
-import {MatFormFieldModule} from '@angular/material/form-field';
-import {MatInputModule} from '@angular/material/input';
-import {MatTooltipModule} from '@angular/material/tooltip';
-import {MatDialogModule} from '@angular/material/dialog';
-import {RequestInterceptorService} from './services/request-interceptor.service';
-import {MatSelectModule} from '@angular/material/select';
-import {MatCheckboxModule} from '@angular/material/checkbox';
-import {MatDatepickerModule} from '@angular/material/datepicker';
-import {MatSlideToggleModule} from '@angular/material/slide-toggle';
-import {DateAdapter, MatNativeDateModule} from '@angular/material/core';
-import {TranslationModule} from './utils/translation/translation.module';
+import {RequestInterceptorService} from './utils/request-interceptor/request-interceptor.service';
+import {DateAdapter} from '@angular/material/core';
 import {AppDateAdapter} from './utils/translation/translation.utils';
-import {SpinnerModule} from './utils/spinner/spinner.module';
-import {SpinnerButtonModule} from './utils/spinner-button/spinner-button.module';
-import {InfoDialogModule} from './utils/info-dialog/info-dialog.module';
 import {NavbarModule} from './navbar/navbar.module';
+import {SpinnerModule} from './utils/spinner/spinner.module';
 
 
 @NgModule({
     declarations: [
         AppComponent,
-        LoginComponent
     ],
     imports: [
         BrowserModule,
         AppRoutingModule,
         BrowserAnimationsModule,
         HttpClientModule,
-        MomentModule,
-        TranslationModule,
-        SpinnerModule,
-        SpinnerButtonModule,
-        InfoDialogModule,
         NavbarModule,
-        MatButtonModule,
-        NgbModule,
-        MatToolbarModule,
-        MatIconModule,
-        MatMenuModule,
-        FlexLayoutModule,
-        MatDividerModule,
-        MatCardModule,
-        FormsModule,
-        MatProgressSpinnerModule,
-        MatFormFieldModule,
-        ReactiveFormsModule,
-        MatInputModule,
-        MatTooltipModule,
-        MatDialogModule,
-        MatSelectModule,
-        MatCheckboxModule,
-        MatDatepickerModule,
-        MatSlideToggleModule,
-        MatNativeDateModule,
+        SpinnerModule
     ],
     providers: [
         {
