+++ angular/src/app/login/login-routing.module.ts
@@ -1,0 +1,14 @@
+import {NgModule} from '@angular/core';
+import {RouterModule, Routes} from '@angular/router';
+import {LoginComponent} from './login.component';
+
+const routes: Routes = [
+    {path: '', component: LoginComponent}
+];
+
+@NgModule({
+    imports: [RouterModule.forChild(routes)],
+    exports: [RouterModule]
+})
+export class LoginRoutingModule {
+}
