+++ angular/src/app/login/login.service.spec.ts
@@ -1,13 +1,13 @@
 import {TestBed} from '@angular/core/testing';
 
-import {UserService} from './user.service';
+import {LoginService} from './login.service';
 
-describe('UserService', () => {
-    let service: UserService;
+describe('LoginService', () => {
+    let service: LoginService;
 
     beforeEach(() => {
         TestBed.configureTestingModule({});
-        service = TestBed.inject(UserService);
+        service = TestBed.inject(LoginService);
     });
 
     it('should be created', () => {
