+++ angular/src/app/utils/request-interceptor/request-interceptor.service.ts
@@ -1,14 +1,16 @@
 import {Injectable} from '@angular/core';
 import {HttpErrorResponse, HttpEvent, HttpHandler, HttpInterceptor, HttpRequest} from '@angular/common/http';
 import {Observable, from} from 'rxjs';
-import {ApiService, LoginStatus} from './api.service';
+import {LoginService, LoginStatus} from '../../login/login.service';
 import {catchError, switchMap} from 'rxjs/operators';
 
 @Injectable()
 export class RequestInterceptorService implements HttpInterceptor {
     constructor(
-        private apiService: ApiService,
-    ) {}
+        private loginService: LoginService,
+    ) {
+        console.log('RequestInterceptor constructor');
+    }
 
     public intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
         if (req.url.indexOf('/auth/') < 0 && req.url.indexOf('/i18n/') < 0) {
@@ -18,10 +20,10 @@ export class RequestInterceptorService implements HttpInterceptor {
             catchError((error: HttpErrorResponse) => {
                 if (error && error.status === 401 && req.url.indexOf('/auth/') < 0) {
                     console.log('Automatically refreshing token');
-                    return from(this.apiService.refreshToken()).pipe(
+                    return from(this.loginService.refreshToken()).pipe(
                         switchMap(() => next.handle(this.addTokenToRequest(req))),
                         catchError(refreshError => {
-                            this.apiService.loginStatusSubject.next(LoginStatus.LOGGED_OUT);
+                            this.loginService.setLoginStatus(LoginStatus.LOGGED_OUT);
                             throw refreshError;
                         })
                     );
@@ -33,7 +35,7 @@ export class RequestInterceptorService implements HttpInterceptor {
 
     private addTokenToRequest(req: HttpRequest<any>): HttpRequest<any> {
         return req.clone({
-            headers: req.headers.set('Authorization', this.apiService.authHeader)
+            headers: req.headers.set('Authorization', this.loginService.authHeader)
         });
     }
 }
